Link site header navbar to template pages that renders based on the link
The url of the page is produced for the link clicked the navbar.
The page h1 of the page is also rendered to match that of the url.
Currently the record that is indexed when the link is clicked is
rendered underneath the hero image this is to be removed before the
deployment.