scripts: Fix carousel displaying incorrect amount of cocktails on resize
Due to when the browser is snapped to a new size the actual resizing
scrolls through all the sizes inbetween. The function I wrote to get
the width of the browser ran to quickly retrieving one of these
inbetween width value and set the number of slides / cocktails based
off of it instead of the final size.